fix:Исправления по результатам второго ревью
