+++ src/main/java/ru/javaadvance/containertracer/controler/dto/CntrDto.java
@@ -2,6 +2,7 @@ package ru.javaadvance.containertracer.controler.dto;
 
 import lombok.Getter;
 import lombok.Setter;
+import ru.javaadvance.containertracer.validators.CntrNumberVal;
 
 
 @Getter
@@ -9,6 +10,7 @@ import lombok.Setter;
 public class CntrDto {
 
     private Long id;
+    @CntrNumberVal(message = "Номер контейнера некоректный")
     private String number;
     private String iso;
     private String size;
