--- src/main/java/ru/javaadvance/containertracer/service/CntrNumberValidator.java
@@ -1,5 +1,0 @@
-package ru.javaadvance.containertracer.service;
-
-public interface CntrNumberValidator {
-    boolean isValid(String cntrNumber);
-}
