+++ src/main/java/ru/javaadvance/containertracer/service/CntrService.java
@@ -8,8 +8,11 @@ import java.util.List;
 
 public interface CntrService {
     List<Cntr> findAll();
+
     Cntr create(Cntr cntr);
+
     void delete(Long id);
+
     void update(Cntr cntr);
 }
